requirements.txt
discord.py>=2.3.2
SQLAlchemy>=1.4
python-dotenv>=1.0
